Update patch set 1
Patch Set 1:

Just warn that new README in dci-openshift-app-agent has been merged to master and, due to that, this CR has merge conflicts with master. It needs to be rebased then. 
Note that the variables table provided in the README has been changed in the new master update, this is the conflict it has.

Patch-set: 1
CC: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>