Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review+1
> > 
> > LGTM but I would wait until having a +1 in CR22333, related to the DCI Openshift Agent https://softwarefactory-project.io/r/c/dci-openstack-agent/+/22333
> 
> There is no link to CR22333 as it is for the OpenStack agent...

yes, true... I thought it was the Openshift agent but it is the OpenStack one. In any case, I was commenting this because that CR was failing in Zuul build and the change is mostly similar to this CR, in terms of the new fields to be added to the dci_job task.

Patch-set: 2
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>