Update patch set 3
Patch Set 3:

It seems this is still a Work-In-Progress but code started very well organized, my only comment for now is to please add a little bit more details in the commit message on next patch, and maybe set the WIP flag in gerrit or add it in the commit message. Thanks!

Patch-set: 3
CC: Gerrit User 908 <908@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>