Generalising registry-related variables in d-o-a-a to make them globally.
Currently, registry variables are only used on the cnf-cert role and are not global. For some new changes, the registry will be needed in more general tasks rather than cnf-cert, so that this change is needed to avoid dealing with different variables pointing to the same content. Pipelines have to be updated too to reflect the new names.

Change-Id: Idd6d7f86866643c8a5282dc191d097908b6dfac7