Update patch set 9
Patch Set 9:

I have been doing some tests by removing lines in get-logs script to avoid saving all the journal logs from the nodes, after discussing with Fred, as this information may not be useful from d-o-a-a perspective. 
As a result, the code has been simplified, only requiring to obtain, from the inventories, the provisioner name and address. It can be checked that it works fine. However, in the last tests, must-gather was not able to retrieve anything from cluster6, comparing this execution to the one from patchset5, but I could see that it is not related to the changes I made in get-logs, as checked in patchet8 test.
So, the code is ready for review.

Patch-set: 9