Update patch set 9
Patch Set 9:

(1 comment)

Ramon, I tested this on SNO[1], here some feedback: Overall looks good, 

I had to add the SSH key of the user that runs the "dci-openshift-app-agent-ctl" command to SSH "provisioner_name" with "provisioner_user", in our dallas labs all that was pre-setup at some point, so I was wondering if we add a note in the README to verify the SSH pub key is in place for this to work.  

This seems to work only in disconnected environments that provide the variables provisionhost_registry*, Is that the type of environments we want to address?

[1] https://www.distributed-ci.io/jobs/ef4ce18f-9fb7-4c30-8361-867ca1c6eda9/jobStates

Patch-set: 9
Reviewer: Gerrit User 908 <908@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>