Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> How likely is it to switch to use the container image instead of building and running the binary? I think that should reduce the amount of steps and dependencies we are dealing with?
> Thoughts?

As I said in Jira, it is not trivial, because example-cnf is currently using tnf v2.0.0 and we had problems with the usage of the containerized version of tnf for that version. I can advance this after using v3.0.0 in example-cnf, but for fixing the current problem in the daily jobs, we need this hotfix for the moment.

Patch-set: 2
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>