Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1: Code-Review+1
> 
> This patch could be tested through an SNO job isn't it?

We can launch an OCP on libvirt or SNO to test d-o-a-a (we will prefer SNO already running to speed thinks up), but I'm not sure this patch will be tested either way. I think this is one of those chicken-and-egg patches hard to test because this .spec builds the rpm, but when this patch is tested the produced RPM is not installed in the system. I might be wrong, I've seen we do a bunch of git clone steps to test the change, but I haven't seen it installs a temporal RPM in a cage or something.

Patch-set: 1
Reviewer: Gerrit User 908 <908@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+1