Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> > Patch Set 1: Code-Review+1
> > 
> > This patch could be tested through an SNO job isn't it?
> 
> We can launch an OCP on libvirt or SNO to test d-o-a-a (we will prefer SNO already running to speed thinks up), but I'm not sure this patch will be tested either way. I think this is one of those chicken-and-egg patches hard to test because this .spec builds the rpm, but when this patch is tested the produced RPM is not installed in the system. I might be wrong, I've seen we do a bunch of git clone steps to test the change, but I haven't seen it installs a temporal RPM in a cage or something.

I fear the same, that running a test will not help. In fact, when testing CR 22648, must-gather works correctly because it was running the code downloaded from gerrit, so in this case, if we launch a job, it will happen the same: the job will be launched from the code downloaded and, as scripts/get-logs is present, it will be executed. 
I think we will have to merge the code, then update dci-openshift-app-agent in the servers to check if it works or not.

Patch-set: 1