Update patch set 1
Patch Set 1:

Hello Sam, as we discussed on IM. Just adding a bit of feedback for your change.

I know it is still WIP but IMHO you may want to review the code created by Ramon/Fred to run the TNF test suite and do some refactor. I think you need to add code to mirror the image, and I do not think an ansible task for each supported test may be easy to maintain. The user should be able to provide a list of tests to execute against the chart, the values file etc. Then you get a report that will be uploaded to DCI.

Please take a look to:
https://github.com/redhat-cip/dci-openshift-app-agent/tree/master/roles

my understanding is that the role that you are working on should be able to test any helm chart if:
- Have the chart files
- Have the proper values for deployment
- Have a cluster with the proper resources
- Have a set of tests to run

In the case of nokia charts, those have multiple dependencies cluster network configurations and relationships with other charts. Maybe when you are done with your module and able to test generic charts we can think on testing one of the core Nokia-chf charts like the one that deploys Keycloak.  If we cannot deploy it on the cluster for a more extensive test maybe just run the general tests.

Patch-set: 1
CC: Gerrit User 724 <724@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>