Update patch set 1
Patch Set 1:

Brandon, so far this looks good, I'll suggest to include the variable in the README, we have a table there with most the variables to use and descriptions.

On the other hand, I have some concerns about leaving the partner deployment, I'll ping you tomorrow to get more insights of your tests with Sudhir.

Patch-set: 1
CC: Gerrit User 908 <908@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>