Update patch set 1
Patch Set 1:

dci-ci-bot job finally passed but it is not testing this feature. I ran a job in dallas and it passed, executing correctly the script. I am about to test it by using dci-check-change, and after this, when confirming it works on libvirt, I will merge this change.

Patch-set: 1