Update patch set 6
Patch Set 6: Code-Review+1

I have also tested it on libvirt (hv7) with the following command, also using the Test-App-Hints and Test-Args-Hints fields in the commit message that you can see:

```
 dci-check-change 22996 <kubeconfig> -e '{ "tnf_targetpodlabels": [environment=test] }'
```

By the way, I had to include the tnf_targetpodlabels list in the CLI command and not in the commit msg because it was failing the way dci-check-change captures Test-Args-Hints for that parameter. I will propose a change to fix it in the near future. As a result of this, dci-ci-bot job will not execute the suite because tnf_targetpodlabels will be null, as it is not present in Test-Args-Hints. So it is not required to wait for dci-ci-bot result (that's why I manually run the test on libvirt with dci-check-change)

The job is here: https://www.distributed-ci.io/jobs/4995b172-d468-4143-8c0c-82df3dba7029/jobStates

If you check it, the test suite is executed correctly, running the access-control tests, and also, the tnf_config.yml file is configured correctly (without using prefix in our case, allowing flexibility in the way the labels are created):

```
---
targetNameSpaces:
  - name: test-check
targetPodLabels:
  - prefix: 
    name: environment
    value: test

certifiedcontainerinfo:

certifiedoperatorinfo:

...
``` 

Regarding Charles' comment, I have included the changes in Mavenir repo in this PR, to be merged when this change is merged: https://github.com/dci-labs/mavenir-tnf-config/pull/2

And as I commented before, it has been tested successfully in dallas too, so the change is ready for review and to be merged if you find it fine

Patch-set: 6
CC: Gerrit User 974 <974@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+1