Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > Just asking: can you provide an example of how to use this variable? Because I don't have clear at all how can it be used
> > I would also include that kind of example in the README, maybe
> 
> It is not to be used directly. dci-check-change will set this variable automatically.

It will do it automatically based on external dependencies of the change to be tested? 
I am saying this because the first idea that came up to my mind after checking this CR is that it would be used to record external repos included in our changes, to create a component related to that repo. Example: if I include a build-depends in my change linking to, e.g., the test-network-function repo, this will create a component for that repo with the commit used. Is it correct?

Patch-set: 7