Update patch set 20
Patch Set 20:

> Patch Set 20:
> 
> > Patch Set 18:
> > ...
> > I'm just having troubles to understand where are you passing those labels in your CI tests. I do not see them in the Args-Hints. So I wonder how are you specifying this.
> 
> I see you are passing the rest of the variables from the pipelines PRs :)

Well, not at all. The thing is, I had a lot of problems by defining a list variable in Test-Args-Hints argument in the commit, so what I did is to pass the tnf_targetpodlabels list when running dci-check-change in hv2 CLI, and there it is working fine. The default libvirt test will not execute the CNF Cert Suite in fact because that parameter is not defined by default in Test-Args-Hints, and that's the workaround I did to test it on hv2. But even not defining the labels, the job will pass on libvirt.

In fact, this is a feature that I want to include in a future change for dci-check-change: to parse correctly complex variables like lists.

Apart from this, the change is working fine, at least for dallas-pipelines PR154 (the last failures mainly come from dallas-pipelines PR138, which is example-cnf update, and I need to review it before merging that PR). And also, there were some jobs failing some hours ago but it was due to a different issue, fixed in CR 23062, or problems with the CNF Cert Suite but not related to d-o-a-a that are under review by CNF Cert Suite tem.

So, as I said, this change is ready to be merged.

Patch-set: 20
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>