Update patch set 21
Patch Set 21:

> Patch Set 20: Code-Review+1
> 
> Ramon, thanks for the clarification. 
> 
> A few more questions:
> - I see in the execution logs the message below[1], as you mentioned the tnf container creates the DaemonSet, but I wonder if this is similar to the partner deployment, where if we do not clean it, we can use it for new tnf test jobs? I would like the tnf scripts/containers will do a better job cleaning the resources they use. but that is not in our control.
> If you think it's ok, we can leave it, or if you think we should add a task in cnf-cert/tasks/post-run.yml to clean it based on a variable, I'm ok too. 
> 
> [1] Error from server (AlreadyExists): error when creating "./temp/debugpartner.yaml": daemonsets.apps "debug" already exists
> 
> - I do not see the variable "tnf_log_level" used anywhere, you added it in the pipelines variables, we can discuss there if you prefer.
> 
> I'll give you a +2 just please can you clarify the above?
> Thanks.

I have included the task for cleaning the debug daemonSet, using the same variable that is used to delete the partner pod (I have renamed it to tnf_postrun_delete_resources), as I have observed that it does not affect the fact that keeping the debug daemonSet created, in the same way than the partner pod, so that we can treat them in the same way. You can see it in the last job I have run, where I set tnf_postrun_delete_resources to false, and the error message [1] appears in execution.log but not affecting to the CNF Cert Suite execution. In the last job that I have run, I set the variable to true to clean the debug daemonSet and it worked, checking on cluster5:

```
[kni@provisioner.cluster5.dfwt5g.lab ~]$ oc get daemonset -n default                                    
No resources found in default namespace.
```

So, unless there is something that I am missing, the change should be ready!

Patch-set: 21
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>