Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Although I think this change is fine and should not break anything from my understanding, I have one question:
> 
> -I don't see the recommendation in the README related to run the post-run hook before the post-run play, can you point me to it? Or if not, could you upload the README to just comment this? Just to justify why this is the only hook that is executed before the corresponding play
> 
> Moreover, I will do a new recheck, because it seems the job has not been scheduled on hv2.

One of our users took the meaning in the post-run hook as the place to copy log files to job_logs.path and they would be uploaded.  I think the readme could be more explicit here and give an example.  I will update this PR to include that update to the readme as well.

Patch-set: 2