Update patch set 4
Patch Set 4:

Hi Sam. Apart from what Fred commented, take care of the new files you have created under the chart-verifier role, because there is some code that is a bare copy of the cnf-cert role, and the idea is that you would have to move what you created in plays/helm_chart_tests into the different tasks of that chart-verifier role (and eventually remove the play, as Fred said).

Patch-set: 4