Update patch set 4
Patch Set 4:

Hi Brandon, I have updated this CR to do the same when we are building the tnf image instead of pulling it from Quay. I have tested both tasks locally on dallas and they worked fine, so the code is correct.

Regarding the fact related to the installation of the collection, I have proposed to include that in this CR: https://softwarefactory-project.io/r/c/dci-openshift-agent/+/22977. So, if you find it fine, maybe we can set this CR as WIP until having that CR merged, so that we ensure that the collection will be installed in the server running the container.

Patch-set: 4
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>