Update patch set 4
Patch Set 4: Workflow-1

Hi all again. I think that, for the moment, we can leave this change opened, because we think this change is more related to following good practices (which are really good) rather than improving what we have (because we are not adding new functionalities; we are just doing the same but in a different way), and it would imply several consequences (e.g. update all the servers in which we have d-o-a-a installed to use this ansible collection).

I have created a Jira issue to review these kind of changes in which we are trying to include good practices in our components, just to define a procedure to be followed in all these possible cases. You can check it here: https://issues.redhat.com/browse/CILAB-337. In the meanwhile, as I said, I think this CR is quite good to improve the way we interact with ansible collections, but for the moment, we would prefer to wait for knowing how to really do these kind of changes.

Patch-set: 4
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Workflow=-1