Update patch set 11
Patch Set 11:

> Patch Set 11: Code-Review+1
> 
> I have some questions here, after checking the full modification that was made from the last change made by Sam:
> 
> -Do we intend, then, to avoid creating a new role for helm chart certification, to have something simple running?

that's just an implementation detail. There is no rule but if the code is simple and doesn't need isolation in a role, having simple tasks is enough.

> -Should we also print the tests that have passed? Because I guess there is no way currently of displaying the content of the claim.json file to do it in the GUI

we'll need to have the discussion with the chart verifier team to rather export in junit.

Patch-set: 11
Reviewer: Gerrit User 20 <20@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>