Update patch set 167
Patch Set 167:

Hello everyone, that you again for the review!

This last change integrates the preflight release 1.0.5 with the workaround for operator-sdk moving tags. Here is a test of all this dev in Dallas as a part of example-cnf pipeline: https://www.distributed-ci.io/jobs/aabd0697-ae4a-418e-bd99-49661e62b394/jobStates  

@Ramon: I've fixed the last typos you mentioned, thank you to notice them, and thank you for the two-step review!
@Beto: The issue with folder cleanup is fixed (thank you for pointing that out!), and I also did some cleanup in Dallas to erase previously missed folders.
@Tony: thank you for the neat ideas to reduce the number of facts by using `json_query`, I used them all :) 
@Jorge: thank you for the refactoring points, I moved runtime-assets and "static" mirroring in the pre-run task, and moved disconnected env things into another separate task as well. I couldn't move cleanup though because I test several operators in a loop and have to clean up after every operator :( 

For information: I had a hope to remove scorecard tests with the preflight 1.0.5 release but some preflight logs are still missing, so let's keep the scorecard for a little bit longer.

Patch-set: 167