Update patch set 6
Patch Set 6:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > I have only tested tnf-test-cnf pipeline correctly, and I have extracted the following conclusions:
> > 
> > -Normal execution where tnf_config is defined correctly - works: https://www.distributed-ci.io/jobs/142e964c-09d2-4f59-bdad-1e65fecc7342/jobStates
> > -If tnf_config is an empty list (tnf_config: []) - the job passes but the CNF Cert Suite results are mostly failures because there are no pods under tests: https://www.distributed-ci.io/jobs/130a150d-6d01-42c7-a31c-748d362721f6/jobStates
> > -If tnf_config is a variable that is not a list (e.g. tnf_config: "") - it fails because loop expects a list and not a bare variable: https://www.distributed-ci.io/jobs/a173271a-59d8-4c4b-8317-6be3c843772a/jobStates#a40f52a6-7ac7-438d-b822-63cef941629d:file5,  but, this means that the checks that I have included (tnf_config is iterable, etc.) are not working. So, do you think it is fine with this result? Or should I change the checks?
> 
> IMO, any case where tnf_config is badly defined a failure should occur. The default value in defaults/main.yml takes care of setting a sane value when is not defined and it should work.

Thanks Tony. I have made two more tests: one using the default value saved in cnf-cert role, and it works (the job passes but CNF Cert Suite tests are not executed at all because there are no pods). The second one is using the pipeline I have defined with the final value for tnf_config in tnf_test_example, but removing the checks related to `tnf_config is iterable` (because `loop` needs a list, so if it is not a list, the job will fail, as we saw in previous tests).

I think it is enough with this, if you find this fine, I think this can be merged.

Patch-set: 6