Update patch set 1
Patch Set 1:

Preflight starts on jumphost which is shared between the clusters. It fails in the case of parallel runs on several clusters (like today, 07/12/2021, cluster5 and cluster6 in parallel) because of the unique preflight container name. The fix is not to fix the name for the preflight container.

Patch-set: 1
Work-in-progress: false