Update patch set 5
Patch Set 5:

(10 comments)

Hi Beto, thank you for the work to structure and uniform the documentation, that should help a lot! I have one generic question - you're basically discarding the large table with all variables listed in one place. I understand that in the current form, it's not readable. But that would be a great source of information-in-one-place if we're looking for something. Maybe we could still somehow have this table in a separate file, just as a list of variables? Wdyt?

Patch-set: 5
Reviewer: Gerrit User 938 <938@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>