Update patch set 5
Patch Set 5:

(5 comments)

> Patch Set 5:
> 
> (10 comments)
> 
> Hi Beto, thank you for the work to structure and uniform the documentation, that should help a lot! I have one generic question - you're basically discarding the large table with all variables listed in one place. I understand that in the current form, it's not readable. But that would be a great source of information-in-one-place if we're looking for something. Maybe we could still somehow have this table in a separate file, just as a list of variables? Wdyt?

Hello Tatiana, yeah. I was thinking on just keeping the variables related to the agent in this table and move all the other variables to the roles documentation. Will allow us to have all in the same place and avoid having to add updates on 2 different places.

I added a row with a reference to the role's readme for the user to check specific details. Let see if the team has other suggestions and alternatives. 

Cheers

Patch-set: 5