Update patch set 8
Patch Set 8:

(21 comments)

Thanks all for your comments. I have updated it accordingly. The rebase didn't help much because the section with the updates was mainly moved to the role's documentation. I made the changes manually and also added the the main readme the section for submitting to Pyxis.

Patch-set: 8