Update patch set 5
Patch Set 5: Code-Review+2

> Patch Set 5:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > > Patch Set 4:
> > > > 
> > > > (1 comment)
> > > > 
> > > > It's fine from my side, but just a minor comment to add, to complement this change :) because we would need some changes in the inventories repo too
> > > 
> > > Hi Ramon, thank you for checking. The variable provision_cache_store is used in d-o-a, so we shouldn't remove it from cluster configs: https://github.com/redhat-cip/dci-openshift-agent/search?q=provision_cache_store 
> > > 
> > > I only simplify the registries credentials definition in d-o-a-a in this change.
> > 
> > Yes, true, but clusterX-post.yml files from inventories are only used in d-o-a-a, not in d-o-a, that's why I commented to update them
> 
> Hi Ramon, I believe that to remove provision_cache_store from inventories is beyond the scope of this change and requires more tests. I only simplify the registries credentials definition in d-o-a-a here. Sorry for the confusion, I refined the description.

Yes, in the end, that change in the inventories is a bit independent from this one, I was double-checking this. I think we can safely merge this one and then update the inventories afterwards (I can do it). So, for the moment, I would say this can be merged and also the PR you opened

Patch-set: 5
Reviewer: Gerrit User 800 <800@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+2