Update patch set 18
Patch Set 18:

(4 comments)

Hello!

Thanks for the feedback. I ended up converting this to a role and added more changes after testing. 
I removed the code that allowed to run single charts. The test suite is able to install charts on a running cluster as part of the tests.

The pipelines used for testing has examples about the different options to install or validate a chart.

As the Core components of the nokia pipelines is not working, I think once this feature of the agent is validated will add it to the Nokia pipeline but will not schedule it as part of dailies. Maybe just reference the nokia pipelines for chart-verifier examples. 

Thanks!

Patch-set: 18
Work-in-progress: false