Remove all_registries_creds def from group_vars/all
group_vars/all default values have more priority than inventory variables. So, if we intend to define all_registries_creds variables only in the inventory, these values are never used because it is taken the default value from group_vars/all, which is "".

This task takes care of removing all_registries_creds from group_vars/all and to define default value in all the occurences of this variable

Change-Id: I6d3a145cd7c1786c77aeb6d45c105182304ad2a7