Fix issues when no DCI job id is used
Some parts of the code are not protected for cases in which DCI job id is not used (e.g. when skipping dci tag); mainly, in the Fail properly tasks, where there is an explicit reference to job_info.job.id.

This change intends to change the message displayed in the error/failure message, depending on job_info.job.id is defined or not.

Change-Id: I45ec55bed40d57c10eec7ac0e698c0a5ab1a486a