Use stable oc client for app-agent
We always download the latest client that is used briefly, later we
downloaded, if not already cached. The same version as the running OCP
cluster. Some functionality in stable versions of oc is lacking in older
versions. Using the downloaded client all the time is a way to get
access to newer functionality, e.g. mirroring FBC catalogs.

We use the stable version of OPM is has been proven succesful so far,
this change is similarly getting the stable version of oc.

Change-Id: Iaa6bf5e90663026dac245f315002a1e084ab62c8