dci-openshift-app-agent-ctl - dev mode should only change config_dir and config if no prefix is provided
In the current logic, if we run d-o-a-a-ctl in dev mode (-d), it is explicitly setting config_dir and config variables to point to the d-o-a-a change. But this is not compatible when using prefixes, because in that case, we would like to use the original CONFIG_DIR as config_dir value to retrieve the correct settings file based on the prefix.

This change allows to only modify config_dir and config in dev mode when we are not using prefixes.

- Currently including fake change to test dci-check-change with doaa too

Depends-On: Ifb7d31ad29453ffb2a05b67264845944746063c2

Change-Id: I5f53521aaa089f5fd72822271c9b9c7318278efb