dci-openshift-app-agent-ctl - dev mode should only change config_dir when nothing in the system
In the current logic, if we run d-o-a-a-ctl in dev mode (-d), it is explicitly setting config_dir and config variables to point to the d-o-a-a change.

This change allows to only modify config_dir and config in dev mode when we run on dev environment with nothing in the system.

- Changing the name of a task in tnf_test_example to make it shorter, and also to see if the change is correctly applied

Depends-On: Ifb7d31ad29453ffb2a05b67264845944746063c2
Change-Id: I5f53521aaa089f5fd72822271c9b9c7318278efb