Add a timeout of 20m to get-logs
In some cases the get-logs has been stuck for several hours, setting a
timeout of 20m to get started should allow to the play to continue.

The extraction of the data in the cluster takes less than a minute,
must-gather is what takes the longes, in the dci-openshift-agent we used
to have a 20m for must-gather and covered most cases.

Change-Id: I5a814f7b703fb20dd18b169149a97788f7ee4756