Creating composite outputs variable
For a pipeline stage to export output data, the "outputs" ansible
variable must be defined as a dictionary, so the keys are set to the
name of the output data used in the pipeline definition and the values
are set to the local path to the data file to be exported.

Currently the agent defines the outputs variable with the path to the
kubeconfig file just before exporting the output files, but this
prevents for custom outputs to be defined in the partner hooks.

With this change, a partner can define the outputs variable with their
custom values in a hook, and this variable will then be combined to the
agent outputs.

Change-Id: Ief8c19b0576ef9a1f7d6eaa89d739e89310fc9f6