verify-tests | Check if fileglob returns any occurence, else fail
In change 29020, the support for fileglobbing in verify-tests was added.

However, this adds an undesired behavior. Prior to this change, if any file to be analyzed by verify-tests was missed, the job failed. Now, since we do the fileglobbing before, what it happens in that case is that the fileglobbed list is a zero-element list, then it does not execute the logic behind parse_junit_file.yml, which has the code that checks if the file exists or not.

We need to check, prior to this, that the fileglobbing is returning anything, else we have to make the job to fail (unless we don't want to have that effect).

Here's an example of a job affected by this issue: https://www.distributed-ci.io/jobs/2c6c7ccb-f997-4f36-8586-1b6f4e3a428e/jobStates?sort=date&task=01d82dd7-1ae6-4ea2-9e8e-b16f6b7c7e77. This job executes cnf-cert, which is not working, so that the XML file with the results is not generated. Fileglobbing returns 0, so that we don't check if the file exists, then the job is not failing, and it should.

Change-Id: If01628669c4d617b83b76d4adaa27a12b16aa235