make the doaa rpm requires python3-openshift
even after installing python3-kubernetes we get an
error of missing openshift module when running doaa.
This change proposes to have python3-openshift as a
requirement rather than an extra step.
python3-openshift was added doa was a dependency of
doaa, but after decoupling the agents we need to
install it, specially for those who only install doaa.

Change-Id: I2284107e5d90dd5ed09b21fd32b184136dec8c4d