Add proper network policies to allow istio-proxy communication in tnf_test_example
Currently, some tnf tests are failing because istio-proxy container is not moving to running status, and it's because of the network policies we are using.
Let's enable traffic between the namespaces under test and istio-system (to reach istio svc) and openshift-dns (to resolve istio svc name).
Also, reordering the location of the network policies' creation, to be done in an earlier stage.

Change-Id: Id4c477186d2d94ca57a8e5755810924796d5dbf4