Fix missing OCP component
OCP components were no longer being created when the app agent was launched. This change fixes the issue.

Additionally, a new test has been added in the hooks to ensure that we can still add the OCP component to an app agent job, making this bug more visible in the future.

TestDallasWorkload: preflight-green

Change-Id: Idb9211063a687f8b65dd52e9abbbdc2b3d3902a7