Include creationTimestamp of event for workload OCP logs
This change request is to replace "LAST SEEN" column with "creationTimestamp" for all k8s events. This change helps in troubleshooting to understand exact timestamp of event creation for Eg: when did pod go to crashloopbackoff or similar issues to analyse exact timestamp of workload pod/container status.

TestDallasWorkload: control-plane-example

Change-Id: I2dfbb99372e002b11e81ab33807c56f50ef49aef